initial commit
The goal is to migrate the existing Neurobi wordpress site over to concise html/css/js hosted on github. Strides have been made but I'm out of time for today. I added most of the content for the bottom two-thirds of the site but still need to add more content, add interactivity, and so on. There was a lot of work done giving the site a nice file structure, and I think all the code is very readable. This is a huge plus!