added overflow handling to body style
this should support responsiveness (or lack thereof) on IOS Safari. I was very pleased with the site on desktop and android mobile, but IOS did not work properly. Specifically, where the screen should cut-off, IOS let users scroll left to right. This also allowed the meet-the-team section (displayed as inline-blocks) to stay stacked left to right instead of responsively stacking vertically. I'd like to eliminate this behavior